Mobile content reordering with Flexbox implementation
A complete page for a site having title, header, navigation bar, a hero image at the main section, other section, footer etc. This have to be responsive in the very small screens of mobile phones. So, to do that we arrange everything in a wrapper element with display flex property.
JQuery used to implement a menu button, so that to toggle display on/off for the navigation buttons. The different structure elements of the page is ordered by numbers 1 to 4 with order property. All of this styles are triggered by a media query for very small screens.