Flexbox single line form implementation
Put every element in a container called cover with flex display capabilities. Give it the value of center to the justify content property. Also give it the value of center to the align items property, to verticaly align. Give the display flex property to the child element, that is a container of the form elements.