Create a mobile app layout with Flexbox implementation
Give to the container element the property of the flex display, and give coloumn at flex direction property. Give scroll value to the overflow property of the content class. Give the flex to the display property of the header too. Also give the header property of align items to be at the center, so that to vertically allign center. Also give the header property of justify content the value of space between.